+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add floating IP support (via 1:1 NAT) in Felix.
+
 ## 1.2.1
 
 - Fix Felix ipset exception when using IPIP.